Extract WarningsFormatter class
To simplify the responsibility of WarningsPanel
